PP-10395 Handle revisiting the success page better
- Previously, if the /register/success page was refreshed or visited
later, an error page was shown because the registration cookie is no
longer present.
- Make it so that the route doesn't require the registration cookie to
be present. If it is present, the `login-after-register` middleware will
log the user in and then the success page will be shown. If it is not
present, the user will either be redirected to the "My services" page if
they are logged in, or the log in page if they are not.

- Inline the middleware that calls the `passport` middleware to log the
user in. Rename the `passport` strategy that it uses to make this easier
to follow.
- Improve the validation for the passport middleware that the
registration cookie is present and add some additional tests.