PP-10395 Show logged in nav on success page
When the user is shown the registration success page, at this point
they have been logged in. Make it so that the navigation bar reflects
this by displaying the links for a logged in user.

Ensure that the  phase banner containing the service navigation and
breadcrumbs is not shown for this page.