BAU Log when a user completes registration
Log a line when a user completes a registration. Include in the log line
the second factor method chosen and whether the user was an internal
GDS user or not.

Include the `internal_user` field for all lines logged for the
registration pages, determining this from the email address used for the
registration.