BAU Log the user external ID for created user
When a user finishes registration, we log a line that includes the
authentication method they picked. Also log the user_external_id of the
created user. This is so we can identify better which types of user are
using an authenticator app if we need to dig into this data.