PP-10099: extend URL validation used for organisation website
Our current method of URL validation has allowed a number of invalid
URLs to pass. This results in the URL being passed to downstream
services which can reject it in an unpredictable way where the error isn't
well handled.

The two outliers:
- spaces in the URL allowing multiple URLs to be enterered (neither
  expected nor wanted)
- incomplete URLs without a TLD specified

The `validatorjs/validator` library includes some fairly gnarly and well
tested regex to ensure that TLD can be required and are well formed.
This library is also actively maintained, quite popular and has no
dependencies.

Include the validator library and use this to catch the exceptional
cases provided above.