PP-10099: add logging tracking to org url validation
We've had report(s) of users confused when entering website addresses as
part of the organisation detail requirements.

Now that we've updated the validation to try and catch more edge cases
in https://github.com/alphagov/pay-selfservice/pull/3704, add some
structured logging to measure how often we are blocking urls as well as
what some common mistakes or misunderstandings might be.