PP-10465 Change where we re-provision the OTP key
Previously we were reprovisioning the OTP key for an invite when:
- The user visited the configure authenticator app page
- The user submitted their phone number

The problem with this was that when the configure authenticator app page
was reloaded due to an invalid security code being submitted, the OTP
key would be regenerated. This would mean the user would have to
configure a new account in their app using the new key, which is not
obvious or user friendly.

Instead, reprovision the code when the user submits the page to choose
their OTP method. This means that the OTP key would be reprovisioned if
the user follows the in-page links to change their method.

There is a small risk that the OTP key wouldn't be provisioned if the
user clicks back in their browser a few times after choosing one method
and then a different method - or if the user visits the URL for
configuring their authenticator app or phone number directly. However,
the repercussions and risk if they were to do this are low.