PP-10805 Send account id when getting agreements
The original intention when agreements were added was to move away from
identifying resources by the gateway account ID, but instead by the
service ID and whether they are live/test.

However, a migration plan hasn't currently been made to move to a world
where there is only one live/test gateway account per service. So for
now, we want to send the account ID when searching for agreements so we
can be sure we're only getting agreements for the current account.

This should be revisited when we can be sure there will only be one
live/test account per service.