PP-10862 Combine all services transactions Cypress tests
Cypress best practices suggest that having tests rely on the state of
previous tests is an anti-pattern.

Combine tests and remove use of Cypress.Cookies.preserveOnce() so that
tests can be run independently.

Remove tests for checking dispute filters. We already have a Cypress
test checking these for the standard list transactions page, and the
functionality being tested is shared.