PP-10862 Combine organisation details Cypress tests
Cypress best practices suggest that having tests rely on the state of
previous tests is an anti-pattern.

Combine tests and remove use of Cypress.Cookies.preserveOnce() so that
tests can be run independently.

Remove the `.within(() => {})` blocks from the tests as they made it
difficult to read once the tests were combined.

Add in `{ delay: 0 }` to the options when calling `.type()` on an
element to speed up text input to make runtime less.