PP-19862 Combine Cypress tests for creating payment links
Cypress best practices suggest that having tests rely on the state of
previous tests is an anti-pattern.

Combine tests and remove use of Cypress.Cookies.preserveOnce() so that
tests can be run independently.

Remove some of the .within(() => {}) blocks from the tests as they made
it difficult to read once the tests were combined.