PP-10862 Make settings Cypress tests standalone
Cypress best practices suggest that having tests rely on the state of
previous tests is an anti-pattern.

Combine tests and remove use of Cypress.Cookies.preserveOnce() so that
tests can be run independently.

Simplify payment-types.cy.js to not attempt to check what the new
selected card types are after updating. This was only checking that the
page has the enabled card types returned by connector selected and not
any selfservice logic. Instead, just check that the page is successfully
reloaded with a success message.