PP-10862 Make switch PSP tests standalone
Cypress best practices suggest that having tests rely on the state of
previous tests is an anti-pattern.

Combine tests and remove use of Cypress.Cookies.preserveOnce() so that
tests can be run independently.

Restructured the tests slightly so that the tests for the individual
pages for submitting the switch details are separate from tests for
the task list logic. I Did this because the task list logic relies purely
on what is returned from connector rather than any client session saved
after a step is completed. This means we don't need to set up new
stubs part way through a test to simulate the task list being updated
after a step is completed.