PP-10863 Use appropriate attributes for email inputs
For all inputs where we expect a full email address, use
type="email" and spellcheck="false" (the latter is recommended by
the GOV.UK Design System for email addresses).

For all other inputs where we expect the user to enter the email
address of themselves or a colleague, use autocomplete="work
email", except for on the sign in and forgotten password pages,
where we use autocomplete="username" because we really want the
username component of the user’s sign-in details to be filled in.

(An argument could be made that we could use
autocomplete="username" on the page where the user enters their
email address when registering. However, it would be more
straightforward to include their email address in a hidden form
field on the page where the user chooses their password to
strongly hint that we want the username and password to be
associated with each other. Therefore, use autocomplete="work
email" on the page where the registering user enters their email
address.)

The email input on the transactions search page is a special case
because it allows partial email addresses. Therefore, both
type="email" and and autocomplete="email" are inappropriate (the
value might not be a full email address), so use type="text" with
no autocomplete attribute. Use inputmode="email" to get a suitable
on-screen keyboard etc., along with spellcheck="false" and
autocapitalize="none" (which is not required for type="email"
because auto-capitalisation should never be applied there).