PP-10887 Add novalidate to all forms
Add a novalidate attribute to all form elements to disable built-in
HTML5 browser client-side validation.

The Design System says to do this at
https://design-system.service.gov.uk/patterns/validation/:

“HTML5 validation is a type of client side validation built into
browsers. Do not use it because:

“• the visual style, placement and content of HTML5 error messages
   cannot be made consistent with the GOV.UK Design System
“• we know that the GOV.UK Design System error message and error
   summary components are accessible

“To turn off HTML5 validation, add ‘novalidate’ to your form tags.”

Strictly speaking, this only needs to be done for forms that
contain validatable fields but it’s easiest just to add it to all
form elements.

Commit 71b1355144a8fcafaa461757ecd551455140ecd3 made this problem
worse by changing several form fields to be type="email" (which
has default browser client-side validation), thus demonstrating
the value of adding novalidate to all form fields.