PP-10863 Transactions filter: attributes for ref and agreement ID
In the filter for searching transactions, apply appropriate
attributes for the reference and agreement ID inputs.

Many on-screen keyboards capitalise the first character of a text
field but this may not be suitable for a reference and is never
suitable for an agreement ID, so set autocapitalize="none".

References are not usually something that can be spell-checked
and agreement IDs can never be spell-checked, so set
spellcheck="false".