PP-10863 Better attributes for 2FA OTP inputs
For inputs where we expect the user to provide a second-factor
authentication one-time password (security code), consistently
use type="text", inputmode="numeric" and pattern="[0+9]+", as
recommended by the GOV.UK Design System for non-decimal numbers.

Ideally we’d use autocomplete="one-time-code" (which can hint where
a one-time password from a text message or authenticator app should
be prefilled) but unfortunately Chrome will remember and suggest
previous values when autocomplete="one-time-code" is used, which is
not what we want since one-time codes are inherently not reusable.
So use autocomplete="off" for now.

Fortunately, a Chromium fix is on the way (see
https://crbug.com/1385892) so hopefully we can change to
autocomplete="one-time-code" in the not-too-distant future.

(Safari and Firefox already seem to be smart enough to not suggest
previous values when autocomplete="one-time-code" is used.)