PP-10834: fix webhooks pagination showing logic
Addresses an issue where the webhooks detail page will choose to not
show pagination if there aren't eneough elements to move to a new page
-- this logic should only stand if we're on the first page, otherwise
"Previous" is a valid move from any non first page.

Correct this and add coverage to check pagination is being returned
appropriately by the service.