PP-11152 Fix loading credentials page
The configure Worldpay credentials page was not loading when no
Worldpay credentials are set. The code was expecting the
`one_off_customer_initiated` key to not exist in the credentials JSON
from connector.

We should change connector so that it doesn't return fields with null
values, but making this change makes the handling of the JSON from
connector a bit more robust.