PP-11120 Spike replacing request with axios
Add axios-base-client.js to replace base.client.js. Replace usage in
connector.client.js as proof of it working.

Added some tests in axios-base-client.test.js, but possibly will need to
add a few more when this is properly implemented.

Have implemented custom retry logic to handle ECONNRESET errors. We
could use the `axios-retry` library for this instead if we don't trust
this.

The intention is that `axios-base-client.js` could be moved to
pay-js-commons. When the client is configure, it accepts an options
object which can provide hook functions to hook into setting the headers
for a request and performing actions on the request starting/ending,
which can be used to do logging per app.

I couldn't get the `transformResponse` callback in the Axios request
config to work, so I've just done the transformation in
`connector.client.js` when handling the response, which is possibly
clearer anyway.