PP-12076 Improve guidance on naming payment links
Update the content on the page for creating a payment link so that
it has guidance on how to name a payment link, and strip back some
of the other text.

New content written by @alanmaddrell, reviewed by @NathanD-GDS and
approved by @Beth-Brown.

Some of the content that was on this page also appears on the page
for managing payment links, shared between them using an include.
Since the content on the two pages now differs, this include has
been deleted and the content that was in it has been directly
incorporated into the page for managing payment links (which has
the exact same content as it did before).