PP-12593 Explain exclusion of Maestro in sandbox accounts
With this change, we are explaining in the UI that Maestro is not available in
SANDBOX accounts.

In TEST accounts and in LIVE accounts, we are showing copy to explain that
3DS needs to be enabled, if it isn't and if it is required.

This is needed because the current hint displayed on production for Maestro,
would link users to a page that is no longer available.

To summarise the scenarios:

- IF (Account is LIVE or TEST) AND (Card requires 3D Secure) AND  (3D Secure is not enabled)

  - THEN The card option is disabled and the hint states: “{card type} cannot be used because 3D Secure is not available. Please contact support.“

- IF (Account is SANDBOX)  AND (Card requires 3D Secure) AND  (3D Secure is not enabled)

  - THEN The card option is disabled and there is a hint stating that “{card type} is not available on sandbox test accounts”.

- OTHERWISE The card option is enabled and there is no hint.

Further information in JIRA.

https://payments-platform.atlassian.net/browse/PP-12593