PP-12632 Add failing test detecting unsanitised content
With this commit, we are adding a failing test showing that an attacker would
be able to make a cross-site scripting attack by entering a script within the
description for an agreement.

The test checks for the visibility of the content of the script tag, which if
it were executed would instead not be visible in the page.

The next commit will fix the failing test by stripping the script tags from
the description field, and thus displaying the content of the attacking
script within the page, thus fulfilling the assertion in the test.

Further information in Jira.

https://payments-platform.atlassian.net/browse/PP-12632