PP-12632 Sanitise description field in agreement page
With the previous commit, we have introduced a failing test showing that an
attacker would be able to make a cross-site scripting attack by entering a
script within the description for an agreement.

With this commit we are now fixing the security vulnerability by stripping the
script tags from the description field, and thus also fixing the failing test
we had just introduced.

Further information in Jira.

https://payments-platform.atlassian.net/browse/PP-12632