PP-13030: Show WORLDPAY TEST SERVICE label
A service is a worldpay test service if it only has one test gateway account
which has payment provider of "worldpay".

This will not break any current services because it used to be that when a
service was created, a sandbox gateway account would be created. We would then
create another worldpay test gateway account (via toolbox) on request. In this
case there would be two test gateway accounts on the service.