BAU — Remove section-* tokens from autocomplete attributes
In some cases, particularly when we make use of partials to put
together a complete form, we mix autocomplete attributes that do
not have a section-* token with attributes that do have a
section-* token. This places them in different autofill scopes,
indicating they do not refer to the same subject, which may hinder
browsers from filling in complete and consistent information.

Remove all the section-* tokens. They are only useful when there’s
a form that asks for information about two different subjects (for
example, two different people) and we do not seem to do that.

Also add a missing autocomplete="organization" attribute.