PP-13622 Fix incorrect value in the admin tool when a dispute is won
With this change, we are fixing a bug for which disputed amounts are shown in
the admin tool always as negative numbers.

The logic will now consider also the state of the dispute and present the
value as positive when the dispute has been won by the service.

Further information in Jira.

https://payments-platform.atlassian.net/browse/PP-13622