PP-14220 Add simplified payment link deletion functionality
With this change, we are adding a new screen in the payment links journey: the
screen is viewed when the user selects ‘Delete’ from the main payment link
screen.

As designed, the user is shown a confirmation page containing just the payment
link name, not the details.

The user is asked to confirm whether they want to delete the payment link.

The user can click on 'Back' or on 'Cancel' and go back to the payment link
index.

The user can click on 'No' and then 'Save changes' and go back to the payment
link index.

If the user clicks on 'Yes' and then 'Save changes', then the service method
to delete the payment link is called.

If the user clicks on 'Save changes', without selecting 'Yes' or 'No', then an
error message is shown, prompting the user to confirm their intent.

This screen is the same for test and live services.

https://payments-platform.atlassian.net/browse/PP-14220