PP-14220 Update controller for the delete payment link functionality
As designed, the user is shown a confirmation page containing just the payment
link name, not the details.

The user is asked to confirm whether they want to delete the payment link.

The user can click on 'Back' or on 'Cancel' and go back to the payment link
index.

The user can click on 'No' and then 'Save changes' and go back to the payment
link index.

If the user clicks on 'Yes' and then 'Save changes', then the service method
to delete the payment link is called.

If the user clicks on 'Save changes', without selecting 'Yes' or 'No', then an
error message is shown, prompting the user to confirm their intent.