PP-14218 PP-14350 Payment link creation steps 1 and 2
With this change, we are implementing the first and second page of the new
payment link creation journey.

In the first page the user can enter the title of the new payment details and
optional details. On clicking continue, we show the second step of the
payment link creation journey, where the user can now decide whether they
want to use an automatic reference generated by GOV.UK or instead use a
custom reference, providing the name for the field and optional hints for the
user.

When clicking `Continue`, we are setting some hardcoded values in the session
for the remaining details and then redirecting to the existing review page
for the old implementation so that the payment link can be created.

The next steps will be:

- create the subsequent pages one by one
- hard code fewer values as we complete the pages
- create the new review page once all the other pages have been created
- redirect to the new index page showing the success banner