PP-14241 Remove superfluous fieldset on transactions page
Remove superfluous and legend-less fieldset for card brand form
controls on transactions page.

On the same page, WAVE complains that the ‘Card brand’ and ‘Payment
status’ label elements are orphans because they reference button
elements after the JavaScript enhances the select to a multi-
select. However, the HTML standard says it is valid for a label
element to reference a button element. Add Nunjucks comments
clarifying this.