PP-14218 Payment Link Creation Reference Page (step 2)
This PR implements the second page of the new payment link creation journey.
In the first page the user can enter the title of the new payment details and
optional details. Previously, on clicking continue, we would redirect to the
review page with hardcoded values. Instead we are now showing the second step
of the payment link creation journey, where the user can now decide whether
they want to use an automatic reference generated by GOV.UK or instead use a
custom reference, providing the name for the field and optional hints for the
user. When clicking `Continue`, we are setting hardcoded values in the
session for the remaining details and then redirecting to the existing review
page for the old implementation so that the payment link can be created.

The next steps will be:

- create the following pages one by one
- hard code fewer values as we complete the pages
- create the new review page once all the other pages have been created
- redirect to the new index page showing the success banner

Steps for review:

* start the new payment link creation journey
* enter a title with or without articles (a, an, the)
* check that the website address is correctly generated (without articles)
* press continue and see the Reference page
* select whether you want an automatic or custom reference
* press Continue and check the old review page
* save the payment link and check that it is created with the correct title,
  details, URL, and also with the desired reference