PP-14438 Improve payment link creation success banner
With this change, we are improving the success banner at payment link
creation, so that it displays an additional message with a link to test the
newly created payment link if the payment link has been created in a test
service or for a live service in sandbox mode.

No message aside from the banner title will be displayed if the service is
live and in normal mode.

Further information in Jira.

https://payments-platform.atlassian.net/browse/PP-14438