PP-14395 Restrict Welsh service name detour to live accounts only
With this change, we are completing the work to prevent the creation of
payment links in Welsh for a service whose name has not been set in Welsh
yet.

With a first PR[1], we have introduced a detour when a service admin starts
the payment link creation journey in Welsh, so that they are directed to the
update service name page for the Welsh option before continuing the payment
link creation.

With this commit, we are now activating this feature only for a service
account in normal mode.

Further information in Jira[2].

[1]
https://github.com/alphagov/pay-selfservice/pull/4713

[2]
https://payments-platform.atlassian.net/browse/PP-14395