PP-14285 Display disputes as negative amounts unless they are 'WON'
With this change, we are improving the display of the amounts associated to
transaction events by correctly displaying disputes as negative amounts.
Unless the event is for a dispute won in favour of the service, in which case
it is correctly displayed as a positive amounts. The new method also includes
refund events which are shown as negative amounts as they were before.

Further information in Jira.

https://payments-platform.atlassian.net/browse/PP-14285